Refactor event components
